endpoint for criteria + npm run resetdb

+++ models/index.js
@@ -18,5 +18,8 @@ module.exports = {
 
 require('./relationship');
 
-db.sequelize.sync({force:true});
-//db.sequelize.sync();
+if (process.argv[2] === 'deleteDb') {
+    db.sequelize.sync({force:true});
+} else {
+    db.sequelize.sync();
+}
